+++ src/card/card-container/card-container.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'ign-card-container',
-  templateUrl: './card-container.html',
-  styleUrls: ['./card-container.scss'],
-})
-export class CardContainerComponent {}