+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### vNEXT
 - Deprecated ign-nav-list, it is recommended to use mat-nav-list instead
 - Deprecated ign-divider, it is recommended to use mat-divider instead
+- Deprecated ign-card-container, it is recommended to use mat-card instead
 - New `ign-list-search` component
 - Added localization to IDS components